+++ lib/dust.dart
@@ -3,15 +3,15 @@
 // BSD-style license that can be found in the LICENSE file.
 
 // Libraries that allow you to create a [Driver] and use it.
-export 'package:dust/src/controller.dart';
 export 'package:dust/src/driver.dart';
-export 'package:dust/src/failure.dart';
+export 'package:dust/src/input_result.dart';
 export 'package:dust/src/failure_library.dart';
-export 'package:dust/src/location.dart';
-export 'package:dust/src/location_canonicalizer.dart';
-export 'package:dust/src/location_scorer.dart';
-export 'package:dust/src/result.dart';
+export 'package:dust/src/path.dart';
+export 'package:dust/src/path_canonicalizer.dart';
+export 'package:dust/src/path_scorer.dart';
 export 'package:dust/src/seed.dart';
 export 'package:dust/src/seed_library.dart';
 export 'package:dust/src/stats.dart';
 export 'package:dust/src/stats_collector.dart';
+export 'package:dust/src/vm_controller.dart';
+export 'package:dust/src/vm_result.dart';