+++ lib/src/path.dart
@@ -10,22 +10,22 @@ import 'package:quiver/core.dart' as q;
 /// integer identifier based on the code offset of the code path. This class
 /// should be constructed with the integer ID unchanged, but the script URI
 /// made shorter & more human readable.
-class Location {
+class Path {
   /// The script URI of the executed code path.
   final String scriptUri;
 
   /// The VM location ID (within a script) for this code path.
   final int locationId;
 
-  /// Construct a [Location] with the given script & VM location ID.
-  Location(this.scriptUri, this.locationId);
+  /// Construct a [Path] with the given script & VM location ID.
+  Path(this.scriptUri, this.locationId);
 
   @override
   int get hashCode => q.hash2(scriptUri.hashCode, locationId.hashCode);
 
   @override
   bool operator ==(Object other) =>
-      other is Location &&
+      other is Path &&
       other.scriptUri == scriptUri &&
       other.locationId == locationId;