+++ lib/src/seed.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:dust/src/result.dart';
+import 'package:dust/src/vm_result.dart';
 
 /// A special input case value which executed new code paths.
 ///
@@ -16,7 +16,7 @@ class Seed {
   double score;
 
   /// The results of the first run of this seed, used to calculate its [score].
-  final Result initialResult;
+  final VmResult initialResult;
 
   /// Create an unscored seed.
   Seed(this.input, this.initialResult);