+++ lib/src/vm_result.dart
@@ -2,24 +2,24 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:dust/src/location.dart';
+import 'package:dust/src/path.dart';
 
 /// Details of how the VM executed a run, including potential error output.
-class Result {
+class VmResult {
   /// Potential error output, if the result was a failure.
   final String errorOutput;
 
   /// The time the VM took during execution.
   final Duration timeElapsed;
 
-  /// The source code locations that the VM executed.
-  final List<Location> locations;
+  /// The source code paths that the VM executed.
+  final List<Path> paths;
 
-  /// Create a [Result] that the VM run succeeded (completed without error).
-  Result(this.timeElapsed, this.locations) : errorOutput = null;
+  /// Create a [VmResult] that the VM run succeeded (completed without error).
+  VmResult(this.timeElapsed, this.paths) : errorOutput = null;
 
-  /// Create a [Result] that the VM run failed, and the error output from that.
-  Result.failed(this.errorOutput, this.timeElapsed, this.locations);
+  /// Create a [VmResult] that the run failed, and the error output from that.
+  VmResult.failed(this.errorOutput, this.timeElapsed, this.paths);
 
   /// Whether the VM run succeeded.
   bool get succeeded => errorOutput == null;